fix corner-case in determining transient bond
when ppar->persist is 1 (suggested in mode 0) and PERCENT is 0.5,
if a bond only last for 1 snapshot, i.e. snap1 = snap2,
it will be labeled as transient, which is not as expected.

The reason is the (int) calculation in checking snapshot interval
left = snap1 - 1 , right = snap1 , while this bond only persist at
one snapshot. As a result, the calculated percent is 0.5